requirements.txt
torch>=2.1.0
torchvision>=0.16.0  
torchaudio>=2.1.0
diffusers>=0.21.4
transformers>=4.33.2
accelerate>=0.22.0
opencv-python>=4.8.1.78
Pillow>=10.0.0
numpy>=1.24.3
scipy>=1.11.2
scikit-image>=0.21.0
timm>=0.9.7
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0
# xformers>=0.0.21  # Opzionale - problemi compilazione su macOS
controlnet-aux>=0.0.7
safetensors>=0.3.3
omegaconf>=2.3.0
einops>=0.6.1
pytorch-lightning>=2.0.7
huggingface_hub>=0.16.4
PyWavelets>=1.4.1
matplotlib>=3.7.2
cog>=0.16.0